catalog: Rehydrated migrated builtin tables
When zero downtime (0dt) is enabled there is a special migrated path
taken for builtin items. Any storage collection that is migrated is
assigned a new empty shard. This commit adds a special case to populate
those new shards with initial data in read-only mode. This is required
to allow downstream dataflows to hydrate.

Works towards resolving #27981