cluster: log disconnects of the GrpcServer
This commit makes sure disconnects of the GRPC connection maintained by
`GrpcServer` are always logged. Previously, there was a log line at the
end of the `ResponseStream` returned by `forward_bidi_stream`, but that
was often not executed on a disconnect because the stream gets droppped
and therefore aborted on disconnects. To always log a message, we need
to instead give the stream a `Drop` impl to do the logging.