catalog: Resolve temp OID collisions
To allocate a unique OID, the durable catalog compares potential OIDs
against all known OIDs. The issue with that approach is that the
durable catalog does not know about any OIDs allocated to a temporary
object, since temporary objects are not stored in the catalog.
Therefore, it's possible to allocate a new OID that was already
allocated to a temporary object.

This commit resolves this issue by requiring callers to pass the
durable catalog a set of existing temporary OIDs before a new one can
be allocated.

Fixes #27719