storage: add metrics for CommandHistory
This commit makes the storage controller's `CommandHistory` track
metrics about the contained command counts. These are mostly useful for
debugging memory leaks caused by the history failing to compact as
expected.

Note that for batched commands (`RunIngestions`, `RunSinks`,
`AllowCompaction`), the metrics count the number of contained entries,
not the number of commands themselves, since this is arguably more
useful. In the future we'll probably want to un-batch these commands
like we have done for compute, which would make the counting less
confusing.