catalog: Implement 0dt builtin item migration (#28124)
This commit implements builtin item migrations that is compatible with
zero down-time upgrades. The issue with the legacy approach is that it
mints new global IDs for each migrated item and its descendents,
without durably writing those IDs down in the catalog. As a result, the
previous Materialize version, which is still running, may allocate the
same global IDs. This would cause confusion for the current version
when it's promoted to the leader because its definition of global IDs
will no longer be valid. At best, the current version would have to
rehydrate all objects that depend on migrated items. At worst, it would
panic.

The high level description of this approach is that we create new
shards for each migrated builtin table with the new table schema,
without changing the global ID. Dependent objects are not modified but
now read from the new shards.

A detailed description of this approach follows. It's important that
all of these steps are idempotent, so that we can safely crash at any
point and non-upgrades turn into a no-op.

   1. Each environment has a dedicated persist shard, called the
      migration shard, that allows environments to durably write down
      metadata while in read-only mode. The shard is a mapping of
      `(GlobalId, deploy_generation)` to `ShardId`.
   2. Collect the `GlobalId` of all migrated tables for the current
      deploy generation.
   3. Read in the current contents of the migration shard.
   4. Collect all the `ShardId`s from the migration shard that are not
      at the current `deploy_generation` or are not in the set of
      migrated tables.
      a. If they ARE NOT mapped to a `GlobalId` in the storage metadata
         then they are shards from an incomplete migration. Finalize
         them and remove them from the migration shard.
         Note: care must be taken to not remove the shard from the
         migration shard until we are sure that they will be finalized,
         otherwise the shard will leak.
     b. If they ARE mapped to a `GlobalId` in the storage metadata then
        they are shards from a complete migration. Remove them from the
        migration shard.
   5. Collect all the `GlobalId`s of tables that are migrated, but not
      in the migration shard for the current deploy generation.
      Generate new `ShardId`s and add them to the migration shard.
   6. At this point the migration shard should only logically contain a
      mapping of migrated table `GlobalId`s to new `ShardId`s for the
      current deploy generation. For each of these `GlobalId`s such
      that the `ShardId` isn't already in the storage metadata:
a. Remove the current `GlobalId` to `ShardId` mapping from the storage
metadata.
      b. Finalize the removed `ShardId`s.
c. Insert the new `GlobalId` to `ShardId` mapping into the storage
metadata.

This approach breaks the abstraction boundary between the catalog and
the storage metadata, but these types of rare, but extremely useful,
abstraction breaks is the exact reason they are co-located.

Since the new shards are created in read-only mode, they will be left
empty and all dependent items will fail to hydrate. Re-hydrating these
tables is left for future work.

Works towards resolving #27981