adapter: when bootstrapping dataflows, pick as_of that is readable
Work towards zero-downtime upgrades: #27406

Before, we were chosing an as_of based on the write frontier of storage
dependencies and the compaction window. Since we don't control the
compaction window at the moment where we determine an as_of this would
lead to an as_of that is not initially readable but only becomes
readable once the next write to depended-upon collections happens.

This starts being problematic when running in read-only mode, during a
0dt upgrade, where we don't write to these storage dependencies and
hence we can never read at the selected as_of.

Now, we pick an as_of that is immediately readable by stepping back from
the upper.