Use SharedRow in Jsonb::from_slice
This uses the shared row to decode jsonb data instead of allocating a new
row. The effect is that we do not need to reallocate a row as we're
decoding. We don't know the size of the decoded row beforehand, so using
the shared row in the limit avoids reallocating.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>