timely-util: fix dataloss bug of ChainBatch::merge_with
The iterator would early exit when encountering the first zero
accumulation, losing all the data contained in the original chain
batches after that point. The code was already filtering for zero
accumulators a couple of lines below so that check was also superfluous.
A test has been added that confirms merging both consolidates and does
not lose data.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>