storage: introduce InstanceMetrics
This commit renames the storage controller's
`RehydratingStorageClientMetrics` to `ReplicaMetrics`, to reflect that
these are metrics tracked per replica. It also introduces a new metrics
type, `InstanceMetrics`, for metrics tracked per instance.

The existing replica metrics are retained but their set of labels
changes from (`instance`) to (`instance_id`, `replica_id`), in line with
the labeling of metrics emitted by the compute controller.

Note that this commit does not attempt to update the
`RehydratingStorageClient`'s usage of the metrics type, leaving the code
in a state that doesn't compile. This is fixed in subsequent commits by
removing the `RehydrationStorageClient`.