storage: add `Instance` and `ReplicaTask` types
This commit introduces the `Instance` controller type, to encapsulate
per-instance controller logic. Initially, this type only takes over
management of the replicas in a storage instance, while the management
of per-instance storage objects is still handled in the top-level
`Controller`.

This commit also introduces a `ReplicaTask` type that handles the
connection to and communication to a replica. It's roughly equivalent to
the `RehydrationTask`, only that it doesn't concern itself with
replaying commands for rehydration, which is handled at the
`Instance`-level.