persist: add a per-shard schema registry
This adds the requirement that schemas handed to persist can be encoded
to and decoded from durably recorded bytes.

It then uses this to assign each schema used with the shard to a unique
identifier. After this PR rolls out and is turned on (validating that at
the moment there is indeed only one in all codepaths), a followup PR
will start tagging batches with the schemas used to encode them. This
will unlock a number of desirable validations in the structured blob
work.

The schema ids are ordered. Once we add persist schema changes, this
means that a persist reader with schema `R` which encounters a batch
written by a writer with schema `W` will be able to directly reason
about what transformations are needed without having to sniff structure
out of the data.

This strategy of directly serializing the schema objects requires that
persist users do the right thing. Specifically, that an encoded schema
doesn't in some later version of mz decode to an in-mem object that acts
differently. In a sense, the current system (before the introduction of
the schema registry) where schemas are passed in unchecked to reader and
writer registration calls also has the same defect, so seems fine.

An alternative is to write down here some persist-specific
representation of the schema (e.g. the arrow DataType). This is a lot
more work and also has the potential to lead down a similar failure mode
to the mz_persist_types `Data` trait, where the boilerplate isn't worth
the safety. Given that we can always migrate later by rehydrating these,
seems fine to start with the easy thing.