storage: Do not advance since in read-only mode
When initializing a storage collection, there is a special case where
if the since of collection is equal to the minimum timestamp, then
we advance the since to the register timestamp. In read-only mode, it's
possible for the timings to line up perfectly where the read-only
instance reads the since of a new collection as the minimum timestamp
and tries to advance the since. The read-only instance should not be
durably modifying the since of any existing storage collections. A live
non-read-only instance may get confused and panic.

This commit fixes the issue by not advancing the since in this special
case in read-only mode.

Works towards resolving #27406