ore: consolidate `DeleteOnDrop` metrics types
This commit cleans up the implementation of the delete-on-drop metrics
by replacing the previous three specialized implementations
(`DeleteOnDropCounter`, `DeleteOnDropGauge`, `DeleteOnDropHistogram`)
with a single `DeleteOnDropMetric type. The previous three types still
exist but are now aliases that fix the `MetricVec` type of the
`DeleteOnDropMetric`.

The result is less redundancy in the implementation of the
delete-on-drop wrappers, and more uniformity on the usage sites.