adapter: detect when a serialized ddl deadlock would occur
In some rare cases the serialized ddl lock could be owned by a session and then
that session tries to execute another statement without clearing its transaction
state (which release the lock). As a workaround, detect this and fail the second
statement instead of deadlocking.

This is not a full fix because we should never be entering this state, but will
suffice as a workaround and unblock the release.

See #28400