adapter,compute,controller: attach workload_class label to relevant metrics
This commit attaches a `workload_class` label whenever a metric is
associated with a workload with a known class (e.g., production,
staging). Relevant metrics are identified as followed:

  * All metrics in a `clusterd` process inherit the workload class of
    the cluster.

  * All metrics in an `environmentd` process that have an `instance_id`
    label inherit the workload class of the identified compute instance
    (cluster).

This will allow us to filter our monitoring based on workload class.

The implementation introduces a new API to the metrics registry called a
"postprocessor", which can mutate the state of all metrics on every call
to `MetricsRegistry::gather`. This was the most efficient implementation
available. In the future, we may want to consider a more explicit
approach of plumbing the `workload_class` label more directly to all
metrics that need it, but that would be an order of magnitude more work
than the approach taken here.

Part of #28097.