txn-wal: fix persist schema hygiene
This closes the "cursed txn-wal TODO" \o/!

Before, we would sometimes fall back to a default "placeholder" schema
in txn-wal (in prod, an empty RelationDesc), but this causes issues with
the soon-to-be-rolled out persist columnar work (see #28146).

Now, we make sure to keep track of data shard write handles that are
handed directly to `register` and use only those for `commit` calls.
This ensures that only a real schema is used to encode batches.

For additional sanity, when applying a batch, use the matching schema by
fetching it from persist if necessary. This fixes issues with background
compaction work spawned as a result of the CaA call.

Closes #28146