persist: re-enable now fixed schema assertion
It turns out we mis-diagnosed the testdrive kafka failures that Parker
was seeing as being caused by the txn-wal TODO. They were actually
caused by shard finalization's placeholder schema, which was fixed in
the #28454 schema registry PR.

This should be enough to get CI green with the desired assertion.
However, since there's a known second occurrence of the bad schema
behavior, keep it as a soft_assert to prevent it from regressing in CI
but also eliminate the possibility for it to cause issues in prod.

Touches #28146