environmentd: don't declare IsLeader before restart
Make it easier to test 0dt deployments by waiting to report the
"IsLeader" deployment status until the deployment is willing to accept
queries as the leader.

Previously, 0dt deployments would report "IsLeader" and *then* restart,
which made it hard for 0dt tests to know when it could start issuing
queries to the deployment after a successful promotion.