transform: fix array access to results (#28538)
While working on #28375, it was noticed that the wrong `Option` was
being unwrapped. `depends` should definitely have the binding
information, but `results` might be missing. The current code
unconditionally tries to access the later, and presumably we never hit
the case where it is missing.

This PR makes sure to use unwrap correctly.