Switch streaming consolidation to an index-based approach
...instead of an iter-based once.

Previously, we had a few different types of iterators that had their own
logic for advancing. Instead, we have a few different types of parts
that can all be indexed uniformly, and we iterate by maintaining
indices into those parts.

This helps simplify things a bit, but the main advantage is that indices
work better with columnar representations, which we're switching to
next.