Use correct types in expression reduction/minimization (#28547)
This corrects a potential bug in `EquivalencePropagation` where it uses
the *output* type of a `Join` for expression minimization. Although the
rough types may be the same, the nullability of the types may change
because of the equivalences in the join, and it is inappropriate to use
that when minimizing the equivalence expressions themselves (we might,
for example, optimize away an `x IS NOT NULL` expression).

At the same time, use the correct input type information for expression
reduction for `Join` and `Reduce`. These operators were rolling back
their "am I a literal" choices because although the expression would be
e.g. the sum of literals, that is not a literal itself. Expression
reduction *requires* type information, even though these reductions may
not, and we make sure to use the correct types here.

### Motivation

There is a potential bug fixed here (not observed in the wild), as well
as an optimization shortfall observed in `join_fusion.slt` that is now
fixed. The latter was related to `Join`, but the same logic is used for
`Reduce`, and it seemed appropriate to fix it at the same time.

### Tips for reviewer

<!--
Leave some tips for your reviewer, like:

    * The diff is much smaller if viewed with whitespace hidden.
    * [Some function/module/file] deserves extra attention.
* [Some function/module/file] is pure code movement and only needs a
skim.

Delete this section if no tips.
-->

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
- <!-- Add release notes here or explicitly state that there are no
user-facing behavior changes. -->