adapter: Allow multi table write transactions
Previously, we disabled write transactions that try to write to more
than one table. The reason was that we had no way of atomically writing
to multiple tables. With the introduction of txn-wal we do have a
method of writing to multiple tables atomically and in fact all writes
in a single group commit are written atomically in a single txn-wal
transaction.

This commit removes the restriction on multi table transactions,
because it is now safe.

Resolves #14567