adapter: Log pgwire connection start
This commit assigns all pgwire connections a unique UUID and logs when
that connection starts in the balancer and adapter. This will help us
debug broken connections. The next step is to log when the connection
is severed, where it was severed and why.

This commit also plumbs dyncfg `ConfigSet`s around so that we can
dynamically turn the logging on and off.

Works towards resolving incidents-and-escalations/#92