mzbuild: get machine-readable Cargo output in prepare_batch
The CargoBuild.prepare_batch function is where we invoke `cargo build`
to actually do the build, so it's also where we should re-invoke with
`--message-format=json` to get the machine readable output.

If we defer the JSON-formatted invocation to CargoBuild.run, we may end
up triggering a recompilation due to different features being enabled,
which can be quite slow.