catalog: Optimize 0dt builtin item migrations (#28612)
This commit optimizes 0dt builtin item migrations by only writing to the
migration shard when there is a migration. Previously, the migrations
would always write to the migration shard, and commit an empty write if
there are no migrations. Additionally, removing the empty write reduces
some contention on the migration shard, for example when a read-only
instance is starting up at the same time as a writeable instance.

Resolves #28605