0dt: Rehydrate migrated introspection sources
When zero downtime (0dt) is enabled there is a special migration path
taken for builtin items. Any storage collection that is migrated is
assigned a new empty shard. This commit adds a special case to populate
those new shards that correspond with introspection sources (A.K.A.
builtin sources). Each of these sources has their own writer task with
a boolean to indicate if it's in read-only mode or not. Usually when in
read-only mode the builtin sources aren't written to. In this commit we
trick migrated builtin sources into thinking that they are not in
read-only mode by setting that bool to `false`. This allows them to
re-hydrate.

We don't have to worry about interference from the old writable
instances, because they don't know about the existence of the new
shards.

Works towards resolving #27981