catalog: Rehydrated migrated builtin tables (#28434)
When zero downtime (0dt) is enabled there is a special migration path
taken for builtin items. Any storage collection that is migrated is
assigned a new empty shard. This commit adds a special case to populate
those new shards with initial data in read-only mode. This is required
to allow downstream dataflows to hydrate.

In order to accomplish this, this commit changes the storage controller
to write to system tables in read-only mode using the legacy
(non-txn-wal) implementation. This is a gross hack, but it is a quick
fix for allowing migrations of builtin tables (see
MaterializeInc#28124). See the comment within the patch for details.

Touches #29781.
Touches MaterializeInc#28124.


---------

Co-authored-by: Nikhil Benesch <nikhil.benesch@gmail.com>