adapter: defer deletes on introspection subscribe reinstalls
This commit changes how introspection subscribes are reinstalled when
the target replica disconnects (but it not dropped). Rather than
immediately clearing the data produced by the subscribe from its target
storage collection, the delete is deferred until a new batch of data
arrives for the reinstalled subscribe.

The intention of this change is to better handle situations were the
target replica needs a long time to reconnect and produce the subscribe
snapshot (e.g. because it is unresponsive). Before this change, there
would be no introspection data for this replica. This could be
confusing, especially since the user is probably not aware of the fact
that the replica disconnected. With this change, the old data is kept
around until there is new data to replace it.