persist: clean up K, V, T, D bounds on Read + Write + Since handles
As the comment mentions, these existed because of the Drop impl. I feel
silly that I didn't think of this technique earlier, but instead of
having the various K, V, T, D, bounds on Drop, we can store the expire
logic as a `BoxFuture<'static, ()>` at handle construction time.

I don't remember the exact history of this, but it's some combination
of:
- Persist was not originally async
- Persist did not originally have auto-expire Drop impls for the handles
- I was still skilling up on async
- I have a tendency to avoid allocs when even remotely possible, even
  when it doesn't matter (as is the case here)

Pure refactor, no behavior change, modulo one alloc when constructing a
read or write handle.