persist: validate that schemas match values in batch construction in CI
This doesn't seem to be happening on the original data write path,
thankfully. However, last week I did end up finding and fixing one known
and one unknown spot where a placeholder schema `RelationDesc::empty`
was used for an empty CaA, which resulted in compaction work being
spawned with the wrong schema. Given that correct schemas are about to
become necessary with the columnar batch work, this will hopefully
prevent regressions.

It's probably too expensive to do this in prod, but seems like a good
idea to catch if any schema shenanigans are happening in CI.