Move `OptimizerMetrics` to `mz_sql`
This is a preparation for using this in the HIR lowering, which is in
`mz_sql`.

This is a somewhat weird place for this. It's due to treating HIR
lowering as part of optimization on the one hand (e.g., being called
from under `optimize/peek.rs`), but having it in `mz_sql`.