catalog: Store fingerprints in 0dt migrations (#28696)
This commit fixes an issue with 0dt builtin item migrations. Materialize
determines if a builtin item is migrated by comparing a compiled in
fingerprint with a durably stored fingerprint. Previously, the 0dt
builtin item migrations did not update the durably stored fingerprint
after migrating an item. This commit fixes that issue.

Resolves #28636