Make `LiteralConstraints` specially handle empty `possible_vals`
This is in the context of robustifying the optimizer.
The earlier state of affairs was a bit spooky, because if some other
transforms didn't properly clean up after `LiteralConstraints`
created an `IndexedFilter` Join with an empty `possible_vals`, then
we'd panic, because the protobuf serialization was assuming that
this field is a non-empty Vec. This panic was never actually
happening, because other transforms always clean up properly, but
better just create an empty constant to make the optimizer more
robust.