persist: hook up decoding from only the structured arrow data
This is done in a way that  relies on the Codec data still being
populated (i.e. for the consolidate optimization above it). Eventually
we'll have to rewrite this path to work entirely without Codec data, but
in the meantime, putting in here allows us to see the performance impact
of decoding entirely from arrow instead of Codec.

Plus, it'll likely be easier to port all the logic here to work solely
on arrow data once we finish migrating things like the
ConsolidatingIter.

Touches #24830