storage/source/postgres: simplify `export_snapshot`
Simplify the `export_snapshot` implementation by fetching the snapshot
name directly from the output of `CREATE_REPLICATION_SLOT`. The prior
implementation used a separate transaction that called
`pg_export_snapshot()`, but this was an extra step that was not
necessary.

The motivation here is to more closely align the approach to
snapshotting PostgreSQL with the forthcoming approach to snapshotting
YugabyteDB.