storage: panic less in production
This commit replaces some panics and asserts in storage code with soft
panics/asserts. It does this for panics related to observing storage
commands/responses and read hold downgrades for collections that do not
exist. These observations indicate bugs in our code and so should be
reported, but continuing afterwards does not compromise correctness, so
we don't want to crash in production.

Note that the storage code contains panics in a lot of other places.
It's likely that many of these should be downgraded to soft panics as
well, but it's not always obvious which ones are essential to uphold
correctness.