compute: make `SequentialHydration::recv` cancel safe
`SequentialHydration::recv` was not cancel safe: Observing received
responses might lead to more dataflows getting scheduled, which requires
sending commands to the replica, but `GenericClient::send` is not
required to be cancel safe.

This is fixed now by introducing a forwarder task that does the direct
interaction with the wrapped client. `SequentialHydration` now only
needs to send commands to a channel connected to that forwarder, which
is non-async and therefore cancel safe.

A slightly annoying bit is that send errors aren't directly returned
from `SequentialHydration::send` anymore, since that call is not waiting
for the send to be successful. Instead they are returned on the next
call to `SequentialHydration::recv`.