adapter: Fix concurrent DDL
Certain DDL statements can trigger an extended DDL transaction, when
run in a multi-statement transaction. These extended DDL
multi-statement transactions have special handling at all stages,
including when they attempt to acquire the DDL lock.

Previously, when executing a statement and acquiring the DDL lock, we
would try and deduce if we are in an extended DDL transaction, by
looking at the type of statement being executed. This is not a robust
check, because those statement that trigger an extended DDL transaction
in multi-statement transactions do not trigger an extended DDL
transaction in single-statement transactions. Currently, the statements
that can trigger an extended DDL transaction are renames and swaps. As
a consequence of this behavior, when a rename or swap are executed in
a single-statement transaction, they never acquire the DDL lock.

This commit fixes this issue, by explicitly checking if we are in an
extended DDL transaction, instead of trying to deduce that state from
the current statement.

Fixes #28640