*: document `GenericClient::recv` impls as cancel safe
The `GenericClient::recv` trait method has documentation that requires
its implementations to be cancel safe. However, it is easy to not be
aware of this when one is making changes to one of the client
implementations, which introduces the risk of `recv` implementations
becoming cancel unsafe. This commit tries to reduce that risk by
explicitly documenting all `GenericClient::recv` impls as cancel safe,
as well as providing justifications in the code.